[fix] restore memory.py
